// src/lib/firefly/model/transaction-split.ts
export enum TransactionSplitTypeEnum {
    Withdrawal = 'withdrawal',
    Deposit = 'deposit',
    Transfer = 'transfer',
    Reconciliation = 'reconciliation',
    OpeningBalance = 'opening balance'
}