// src/lib/firefly/model/transaction-type-filter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Firefly III API v1.5.4
 * This is the documentation of the Firefly III API. You can find accompanying documentation on the website of Firefly III itself (see below). Please report any bugs or issues. You may use the \"Authorize\" button to try the API below. This file was last generated on 2021-09-25T14:21:28+00:00 
 *
 * The version of the OpenAPI document: 1.5.4
 * Contact: james@firefly-iii.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @enum {string}
 */

export enum TransactionTypeFilter {
    All = 'all',
    Withdrawal = 'withdrawal',
    Withdrawals = 'withdrawals',
    Expense = 'expense',
    Deposit = 'deposit',
    Deposits = 'deposits',
    Income = 'income',
    Transfer = 'transfer',
    Transfers = 'transfers',
    OpeningBalance = 'opening_balance',
    Reconciliation = 'reconciliation',
    Special = 'special',
    Specials = 'specials',
    Default = 'default'
}